Adds initial configuration, connection, and objectStore support.
This change fleshs out initConfig() and init(), adding configuration
properties for the FileNet Content Engine URL, Object Store name,
and access credentials.

This change imports the Proxies and ProxyAdaptorContext classes
from the Documentum adaptor, unchanged except for the package.

The ObjectFactory class is similar to that in the v3 connector
and is expected to add two additional methods for getting
ProperyFilters and SearchWrapper.

FileNet maintains its access credentials in thread-local
storage, and they must be pushed and popped in matching
pairs to avoid resource leakage. The Connection class
wraps the FileNet Connection, with an AutoClosable implentation
that ensures the UserContext get popped when we are finished
with the Connection.

Code Review: http://codereview.appspot.com/266900043